refactor(convert): extract replaced module with try_convert entry
Phase 1 of fulgur-odgo. Move 6 image/svg/content-url helpers from
convert/mod.rs into convert/replaced.rs:

- pub(super): make_image_pageable, resolve_image_dimensions
- private: convert_image, convert_svg, convert_content_url,
  wrap_replaced_in_block_style

Add pub(super) fn try_convert(doc, node_id, ctx) -> Option<...> that
wraps the dispatcher's three replaced branches into a single entry
point. Preserves original branch order and fall-through semantics.

pseudo.rs gains an explicit use super::replaced::{...} for its two
cross-module calls; mod tests block updated to super::replaced::
prefix for make_image_pageable.

No behavioral change. VRT byte-identical against committed goldens.
