+++ crates/fulgur/src/convert/pseudo.rs
@@ -2,6 +2,7 @@ use super::positioned::{
     AbsCb, build_absolute_pseudo_children, is_absolutely_positioned,
     try_build_absolute_pseudo_image,
 };
+use super::replaced::{make_image_pageable, resolve_image_dimensions};
 use super::*;
 
 /// Build an `ImagePageable` for a `::before`/`::after` pseudo-element node
